requirements.txt
flask~=2.3.3
requests~=2.32.3
beautifulsoup4~=4.12.3
numpy~=1.26.4
python-dotenv~=1.0.1
pylatex~=1.4.2
bs4~=0.0.2
